chore: remove/move other mislocated items
